+++ env.d/development/common.e2e.dist
@@ -1,7 +1,6 @@
 # For the CI job test-e2e
-SUSTAINED_THROTTLE_RATES="200/hour"
 BURST_THROTTLE_RATES="200/minute"
 DJANGO_SERVER_TO_SERVER_API_TOKENS=test-e2e
+SUSTAINED_THROTTLE_RATES="200/hour"
 Y_PROVIDER_API_KEY=yprovider-api-key
 Y_PROVIDER_API_BASE_URL=http://y-provider:4444/api/
-THEME_CUSTOMIZATION_FILE_PATH="" #force theme_customization to be empty
