+++ src/frontend/apps/e2e/__tests__/app-impress/home.spec.ts
@@ -1,6 +1,6 @@
 import { expect, test } from '@playwright/test';
 
-import { CONFIG } from './common';
+import { overrideConfig } from './common';
 
 test.beforeEach(async ({ page }) => {
   await page.goto('/docs/');
@@ -54,18 +54,8 @@ test.describe('Home page', () => {
   });
 
   test('it checks the homepage feature flag', async ({ page }) => {
-    await page.route('**/api/v1.0/config/', async (route) => {
-      const request = route.request();
-      if (request.method().includes('GET')) {
-        await route.fulfill({
-          json: {
-            ...CONFIG,
-            FRONTEND_HOMEPAGE_FEATURE_ENABLED: false,
-          },
-        });
-      } else {
-        await route.continue();
-      }
+    await overrideConfig(page, {
+      FRONTEND_HOMEPAGE_FEATURE_ENABLED: false,
     });
 
     await page.goto('/');
